requirements.txt
aioredis==1.1.0
asgiref==2.3.2
astroid==2.0.2
async-timeout==3.0.0
attrs==18.2.0
autobahn==18.9.2
Automat==0.7.0
autopep8==1.4.1
certifi==2018.4.16
channels==2.1.3
channels-redis==2.3.0
chardet==3.0.4
constantly==15.1.0
daphne==2.2.2
Django==2.1
django-filter==2.0.0
django-silk==3.0.1
djangorestframework==3.8.2
gprof2dot==2016.10.13
hiredis==0.2.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
isort==4.3.4
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
msgpack==0.5.6
mysqlclient==1.3.12
pycodestyle==2.4.0
Pygments==2.2.0
PyHamcrest==1.9.0
pylint==2.1.1
pylint-django==2.0
pylint-plugin-utils==0.4
python-dateutil==2.7.3
python-http-client==3.1.0
pytz==2018.5
requests==2.19.1
sendgrid==5.6.0
six==1.11.0
sqlparse==0.2.4
Twisted==18.7.0
txaio==18.8.1
typed-ast==1.1.0
typing==3.6.4
urllib3==1.23
wrapt==1.10.11
zope.interface==4.5.0